requirement.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cli-helpers==2.1.0
click==7.1.2
configobj==5.0.6
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.5
django-allauth==0.44.0
django-filter==2.4.0
django-rest==0.8.6
django-rest-auth==0.9.5
djangorestframework==3.12.2
idna==2.10
importlib-metadata==3.4.0
Markdown==3.3.3
mycli==1.22.2
oauthlib==3.1.0
prompt-toolkit==3.0.10
pycparser==2.20
Pygments==2.7.3
PyJWT==2.0.0
PyMySQL==1.0.2
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
tabulate==0.8.7
terminaltables==3.1.0
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
zipp==3.4.0
